+++ app.json
@@ -1,6 +1,6 @@
 {
   "expo": {
-    "name": "Ecoh",
+    "name": "ecoh",
     "slug": "Ecoh",
     "version": "1.0.0",
     "orientation": "portrait",
@@ -9,7 +9,8 @@
     "userInterfaceStyle": "automatic",
     "newArchEnabled": true,
     "ios": {
-      "supportsTablet": true
+      "supportsTablet": false,
+      "bundleIdentifier": "Anthony.McMillan.ecoh"
     },
     "android": {
       "adaptiveIcon": {
@@ -24,6 +25,7 @@
     },
     "plugins": [
       "expo-router",
+      "expo-video",
       [
         "expo-splash-screen",
         {
@@ -36,6 +38,14 @@
     ],
     "experiments": {
       "typedRoutes": true
+    },
+    "extra": {
+      "router": {
+        "origin": false
+      },
+      "eas": {
+        "projectId": "38fc3634-a451-43e0-94e1-a5918e340d58"
+      }
     }
   }
 }
