+++ app/(tabs)/dashboard/index.tsx
