+++ app/(tabs)/profile/index.tsx
@@ -5,7 +5,7 @@ const ProfileIndex = () => {
   const router = useRouter();
 
   useEffect(() => {
-    router.replace('/(authenticated)/(tabs)/profile/posts'); // Redirect to 'posts' as the default tab
+    router.push('/profile/posts'); // Redirect to 'posts' as the default tab
   }, []);
 
   return null; // Render nothing while redirecting
