+++ app/(tabs)/profile/tags.tsx
@@ -1,15 +1,20 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { View, Text, StyleSheet } from 'react-native';
 import { useTheme } from '@/src/theme/ThemeContext';
 import { typography } from '@/src/theme/typography';
 
-const PostsScreen: React.FC = () => {
+const TagsScreen: React.FC = () => {
   const { colors } = useTheme();
+
+  useEffect(() => {
+    console.log('TagsScreen Mounted');
+    return () => console.log('TagsScreen Unmounted');
+  }, []);
 
   return (
     <View style={[styles.container, { backgroundColor: colors.background }]}>
       <Text style={[styles.text, { color: colors.text }]}>
-        Posts content goes here.
+        Tagged content goes here.
       </Text>
     </View>
   );
@@ -26,4 +31,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default PostsScreen;
+export default TagsScreen;
