+++ app/(tabs)/search/index.tsx
