+++ package.json
@@ -16,22 +16,23 @@
   },
   "dependencies": {
     "@expo-google-fonts/poppins": "^0.2.3",
-    "@expo-google-fonts/roboto": "^0.2.3",
     "@expo/vector-icons": "^14.0.2",
     "@react-native-async-storage/async-storage": "^2.0.0",
-    "@react-native-community/slider": "^4.5.5",
+    "@react-native-community/datetimepicker": "^8.2.0",
     "@react-native-masked-view/masked-view": "^0.3.2",
     "@react-navigation/bottom-tabs": "^7.0.0",
+    "@react-navigation/material-top-tabs": "^7.0.18",
     "@react-navigation/native": "^7.0.0",
+    "@shopify/flash-list": "^1.7.2",
     "@tanstack/query-async-storage-persister": "^5.60.6",
     "@tanstack/react-query": "^5.60.6",
     "@tanstack/react-query-persist-client": "^5.60.6",
     "axios": "^1.7.7",
     "dayjs": "^1.11.13",
     "expo": "~52.0.4",
-    "expo-av": "^15.0.1",
     "expo-blur": "~14.0.1",
     "expo-constants": "~17.0.2",
+    "expo-dev-client": "~5.0.4",
     "expo-fast-image": "^1.1.3",
     "expo-font": "~13.0.1",
     "expo-haptics": "~14.0.0",
@@ -42,20 +43,14 @@
     "expo-secure-store": "^14.0.0",
     "expo-splash-screen": "~0.29.7",
     "expo-status-bar": "~2.0.0",
-    "expo-symbols": "~0.2.0",
-    "expo-system-ui": "~4.0.2",
-    "expo-web-browser": "~14.0.0",
+    "expo-video": "^2.0.1",
     "formik": "^2.4.6",
-    "i18n-js": "^4.5.0",
     "libphonenumber-js": "^1.11.14",
     "lodash": "^4.17.21",
-    "lottie-ios": "^4.5.0",
-    "lottie-react-native": "^7.1.0",
     "moment": "^2.30.1",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-native": "0.76.1",
-    "react-native-awesome-slider": "^2.5.6",
     "react-native-country-codes-picker": "^2.3.5",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-modal-datetime-picker": "^18.0.0",
@@ -66,7 +61,6 @@
     "react-native-screens": "~4.0.0",
     "react-native-svg": "^15.9.0",
     "react-native-svg-transformer": "^1.5.0",
-    "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.2",
     "styled-components": "^6.1.13",
     "yup": "^1.4.0",
@@ -77,6 +71,7 @@
     "@types/jest": "^29.5.12",
     "@types/lodash": "^4.17.13",
     "@types/react": "~18.3.12",
+    "@types/react-native": "^0.73.0",
     "@types/react-test-renderer": "^18.3.0",
     "jest": "^29.2.1",
     "jest-expo": "~52.0.0",
