+++ src/components/post/Badge.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Text, StyleSheet, ViewStyle, TextStyle, View } from 'react-native';
+import { Text, ViewStyle, TextStyle, View } from 'react-native';
 import { LinearGradient } from 'expo-linear-gradient'; // Import LinearGradient from Expo
 import { typography } from '@/src/theme/typography';
 
@@ -31,18 +31,16 @@ const Badge: React.FC<BadgeProps> = ({ type }) => {
 
   if (type === 'connection') {
     return (
-      <View style={styles.neuomorphicContainer}>
-        <LinearGradient
-          colors={['#00c6ff', '#0072ff']}
-          style={badgeContainer}
-          start={{ x: 0, y: 0 }}
-          end={{ x: 1, y: 0 }}
-          accessible
-          accessibilityLabel={badgeText}
-        >
-          <Text style={textStyle}>{badgeText}</Text>
-        </LinearGradient>
-      </View>
+      <LinearGradient
+        colors={['#00c6ff', '#0072ff']}
+        style={badgeContainer}
+        start={{ x: 0, y: 0 }}
+        end={{ x: 1, y: 0 }}
+        accessible
+        accessibilityLabel={badgeText}
+      >
+        <Text style={textStyle}>{badgeText}</Text>
+      </LinearGradient>
     );
   } else {
     return (
@@ -65,17 +63,5 @@ const Badge: React.FC<BadgeProps> = ({ type }) => {
     );
   }
 };
-
-const styles = StyleSheet.create({
-  neuomorphicContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    shadowColor: '#888',
-    shadowOffset: { width: 0, height: 4 }, // Downward shadow for depth
-    shadowOpacity: 0.25,
-    shadowRadius: 4, // Soft spread for shadow
-    elevation: 8, // Android shadow effect
-  },
-});
 
 export default Badge;
