+++ src/types/navigation.ts
@@ -1,0 +1,6 @@
+export type ProfileTabParamList = {
+  Posts: undefined;
+  Media: undefined;
+  Tags: undefined;
+  Saved: undefined;
+};
